refactor(core): organize API structure
* Created folders for api, classes, interfaces and types to unclutter the files.
