+++ templates/wortal-api/api/Ads.ts
@@ -1,3 +1,6 @@
+import { PlacementType } from "../types/Ads";
+import { ErrorMessage} from "../interfaces/Wortal";
+
 /**
  * Shows an interstitial ad. These can be shown at various points in the game such as a level end, restart or a timed
  * interval in games with longer levels.
@@ -34,9 +37,11 @@ export function showInterstitial(placement: PlacementType,
  * must be notified of the ad and give permission to show before it can be shown.
  * @example
  * // This example shows the game flow independent of the outcome of the ad.
+ * // Ex: Player gets bonus coins for watching the ad, but the game continues regardless of the outcome.
  * Wortal.ads.showRewarded('BonusCoins', pauseGame, resumeGame, skipBonus, addBonusCoins);
  *
  * // This example shows the game flow depending on the outcome of the ad.
+ * // Ex: Player dies and can revive by watching an ad, but if they skip the ad they lose the level.
  * Wortal.ads.showRewarded('ReviveAndContinue', pauseAudio, resumeAudio, endGame, continueGame);
  * @param description Description of the placement.
  * @param beforeAd Callback for before the ad is shown. Pause the game here.
@@ -58,21 +63,3 @@ export function showRewarded(description: string,
                              noFill?: Function): void {
     (window as any).Wortal.ads.showRewarded(description, beforeAd, afterAd, adDismissed, adViewed, noFill);
 }
-
-/**
- * Types of ad placements as defined by Google:
- * https://developers.google.com/ad-placement/docs/placement-types
- * <pre>
- *     start: Your game has loaded, the UI is visible and sound is enabled, the player can interact with the game,
- *     but the game play has not started yet.
- *
- *     pause: The player pauses the game.
- *
- *     next: The player navigates to the next level.
- *
- *     browse: The player explores options outside of gameplay.
- * </pre>
- * NOTE: preroll and reward are reserved for special ad calls. Do not pass these types to any ads API calls.
- * Use ads.showRewarded() which will automatically add the 'reward' type.
- */
-export type PlacementType = 'preroll' | 'start' | 'pause' | 'next' | 'browse' | 'reward'
