+++ templates/wortal-api/api/Player.ts
@@ -1,3 +1,7 @@
+import { WortalPlayer } from "../classes/WortalPlayer";
+import { ConnectedPlayerPayload, SignedASID } from "../interfaces/Player";
+import { ErrorMessage} from "../interfaces/Wortal";
+
 /**
  * Gets the player's ID from the platform.
  * @example
@@ -126,7 +130,7 @@ export function flushDataAsync(): Promise<void> {
  *     hoursSinceInvitation: 4,
  * }).then(players => console.log(players.length);
  * @param payload Options for the friends to get.
- * @returns {Promise<ConnectedPlayer[]>} A promise that resolves with a list of connected player objects.
+ * @returns {Promise<WortalPlayer[]>} A promise that resolves with a list of connected player objects.
  * @throws {ErrorMessage} See error.message for details.
  * <ul>
  * <li>NOT_SUPPORTED</li>
@@ -177,7 +181,7 @@ export function getSignedPlayerInfoAsync(): Promise<object> {
  * <li>RETHROW_FROM_PLATFORM</li>
  * </ul>
  */
-export function getASIDAAsync(): Promise<string> {
+export function getASIDAsync(): Promise<string> {
     return (window as any).Wortal.player.getASIDAAsync();
 }
 
@@ -242,77 +246,4 @@ export function canSubscribeBotAsync(): Promise<boolean> {
  */
 export function subscribeBotAsync(): Promise<void> {
     return (window as any).Wortal.player.subscribeBotAsync();
-}
-
-/**
- * Represents a single player.
- */
-export class WortalPlayer {
-    private _current: WortalPlayerData;
-    constructor(player: WortalPlayerData) {
-        this._current.id = player.id;
-        this._current.name = player.name;
-        this._current.photo = player.photo;
-    }
-
-    get id(): string {
-        return this._current.id;
-    }
-
-    get name(): string {
-        return this._current.name;
-    }
-
-    get photo(): string {
-        return this._current.photo;
-    }
-}
-
-/**
- * Payload options for players.
- */
-export interface ConnectedPlayerPayload {
-    cursor?: number;
-    filter?: ConnectedPlayerFilter;
-    hoursSinceInvitation?: number;
-    size?: number;
-}
-
-/**
- * Data about a player.
- */
-export interface WortalPlayerData {
-    id: string;
-    name: string;
-    photo: string;
-}
-
-/**
- * Filters that can be applied to connected player searches.
- */
-export type ConnectedPlayerFilter = 'ALL' | 'INCLUDE_PLAYERS' | 'INCLUDE_NON_PLAYERS' | 'NEW_INVITATIONS_ONLY'
-
-/**
- * Represents app-scoped user id of current player along with a signature to verify that it indeed comes from Facebook.
- */
-export interface SignedASID {
-    /**
-     * The ID of the player
-     */
-    asid: string;
-    /**
-     * A signature to verify this object indeed comes from Facebook. The string is base64url encoded and signed with an
-     * HMAC version of your App Secret, based on the OAuth 2.0 spec.
-     *
-     * You can validate it with the following 4 steps:
-     *
-     * - Split the signature into two parts delimited by the '.' character.
-     * - Decode the first part (the encoded signature) with base64url encoding.
-     * - Decode the second part (the response payload) with base64url encoding, which should be a string representation of a JSON object that has the following fields: ** algorithm - always equals to HMAC-SHA256 ** issued_at - a unix timestamp of when this response was issued. ** asid - the app-scoped user id of the player.
-     * - Hash the whole response payload string using HMAC SHA-256 and your app secret and confirm that it is equal to the encoded signature.
-     * - You may also wish to validate the issued_at timestamp in the response payload to ensure the request was made recently.
-     *
-     * Signature validation should only happen on your server. Never do it on the client side as it will compromise your app secret key.
-     */
-    signature: string;
 }
