+++ templates/wortal-api/interfaces/Session.ts
@@ -1,0 +1,7 @@
+/**
+ * Data about the traffic source for this game.
+ */
+export interface TrafficSource {
+    ['utm_source']?: string;
+    ['r_entrypoint']?: string;
+}
