+++ templates/wortal-api/types/Ads.ts
@@ -1,0 +1,17 @@
+/**
+ * Types of ad placements as defined by Google:
+ * https://developers.google.com/ad-placement/docs/placement-types
+ * <pre>
+ *     start: Your game has loaded, the UI is visible and sound is enabled, the player can interact with the game,
+ *     but the game play has not started yet.
+ *
+ *     pause: The player pauses the game.
+ *
+ *     next: The player navigates to the next level.
+ *
+ *     browse: The player explores options outside of gameplay.
+ * </pre>
+ * NOTE: preroll and reward are reserved for special ad calls. Do not pass these types to any ads API calls.
+ * Use ads.showRewarded() which will automatically add the 'reward' type.
+ */
+export type PlacementType = 'preroll' | 'start' | 'pause' | 'next' | 'browse' | 'reward'
