+++ templates/wortal-api/types/Player.ts
@@ -1,0 +1,4 @@
+/**
+ * Filters that can be applied to connected player searches.
+ */
+export type ConnectedPlayerFilter = 'ALL' | 'INCLUDE_PLAYERS' | 'INCLUDE_NON_PLAYERS' | 'NEW_INVITATIONS_ONLY'
