+++ templates/wortal-api/types/Session.ts
@@ -1,0 +1,4 @@
+/**
+ * Platform the game is currently being played on.
+ */
+export type Platform = 'wortal' | 'link' | 'viber' | 'gd' | 'facebook' | 'debug'
